refactor: user | move auth function to utils for reuse
