+++ src/utils/isValidEmail.js
@@ -1,0 +1,7 @@
+// Check email or phone number
+const isValidEmail = (email) => {
+  const regex = /^[a-zA-Z0-9._-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/;
+  return regex.test(email);
+};
+
+export default isValidEmail;
